Moving MGOpt into the main source tree
